综述: 优化数据类型，清理冗余代码与日志，并改进 NoticeReadLimiter 实现
- 将 `WriteListener` 及其相关实现中的 `long` 类型参数改为 `int`，以优化内存和性能。
- 清理 `DefaultPipeline`、`HttpConnection` 和 `ProxyHttpHandler` 中的 `uid`、`pipelineUid` 及相关的 `TRACEID` 使用，移除冗余的 `toString` 方法。
- 移除多处日志输出语句和注释掉的调试代码，简化代码逻辑。
- 重构 `NoticeReadLimiter` 的并发控制逻辑，增加 `readCompleted` 方法，提升读取限流的准确性。
- 对 `HttpUrl` 和 `ProxyHttpHandler` 进行代码格式化和结构优化。

+++ src/main/java/cc/jfire/jnet/common/api/WriteListener.java
@@ -5,12 +5,12 @@ public interface WriteListener
     WriteListener INSTANCE = new WriteListener()
     {
         @Override
-        public void partWriteFinish(long currentSend)
+        public void partWriteFinish(int currentSend)
         {
         }
 
         @Override
-        public void queuedWrite(long size)
+        public void queuedWrite(int size)
         {
         }
 
@@ -23,14 +23,14 @@ public interface WriteListener
     /**
      * @param currentSend 本次已经写出的容量
      */
-     void partWriteFinish(long currentSend) ;
+     void partWriteFinish(int currentSend) ;
 
     /**
      * 当前入队的待写出大小
      *
      * @param size
      */
-     void queuedWrite(long size) ;
+     void queuedWrite(int size) ;
 
      void writeFailed(Throwable e);
 }
