+++ src/main/java/cc/jfire/jnet/common/internal/DefaultWriteCompleteHandler.java
@@ -28,7 +28,7 @@ public class DefaultWriteCompleteHandler extends AbstractWriteCompleteHandler im
                 socketChannel.write(byteBuffer, byteBuffer, this);
                 return;
             }
-            long currentSend = sendingData.getWritePosi();
+            int currentSend = sendingData.getWritePosi();
             writeListener.partWriteFinish(currentSend);
             sendingData.clear();
             if (!queue.isEmpty())
